+++ src/components/TransactionsTracker/TransactionTracker.tsx
@@ -1,5 +1,5 @@
 import { useBatchTransactionsTracker } from 'hooks/transactions/batch/tracker/useBatchTransactionsTracker';
-import { useTransactionsTracker } from 'hooks/transactions/tracker/useTransactionsTracker';
+import { useTransactionsTracker } from 'hooks/transactions/useTransactionsTracker/useTransactionsTracker';
 import { TransactionsTrackerType } from 'types/transactionsTracker.types';
 
 export function TransactionsTracker(props: TransactionsTrackerType) {
